docall should really be provided by lua54.lua, not api.lua
The NeXUS API should be in api.lua, since it needs the added state of 
the VM table in order to function. The facilities that allow for that 
added state also need to be in api.lua, since that's where they're used. 
Something as simple as a docall function should probably be in 
lua54.lua.